RAG Model with Gemini/requirements.txt
streamlit
langchain-community
langchain
langchain-google-genai
langchain-chroma
chromadb
python-dotenv
PyPDF2
tiktoken
requests